Re-arrange file/function structure
