+++ credentials.py
@@ -1,0 +1,8 @@
+#Details from Reddit 'App creation' page
+CLIENT_ID = 'UJDxRErWoo4Sjg'
+CLIENT_SECRET = 'knnySEplo-Aarpg9fJH2BRrgCw0'
+#Reddit login credentials
+USER_NAME = 'EtherA65'
+USER_PASSWORD = 'eThA23'
+#Used to identfy app (Firefox, Chrome, etc.)
+USER_AGENT = 'Hedone_v1'
